SDL_ttf is now initted in the Font class.
Added a static variable to the Font class to keep track of the amount of fonts that are actually opened with automatic incrementation and decrementation. Also created a function to return the amount of open fonts. SDL_ttf's initialization is now piggybacked onto the Font class, if the API wasn't initialized before, it is now done in the constructor and at the destructor, if it is the last instance to be using fonts, TTF_Quit() is now called after closing the last font.