Modified Font class constructor, added a function.
Removed responsibility of SDL_ttf from the Font class to the programmer
using the library because initialization of SDL_ttf will not work at that
scope and because of design decisions. Where the initialization was is now
a simple check and warning if it isn't initialized. Added a function to
allow the programmer to close the font file themselves with a simple call
to closeFont(). Also replaced sequence of code to remove a font in the
loadFont function with a simple call to the new closeFont() function. The
Font class is now less verbose.