Modified StateManager and Window class functions.
Removed unnecessary result boolean from pushState in the StateManager class
and put the state initialization call into the if statement. In the Window
class, all drawing functions now call clear(). clear() handles clearing the
screen with the default color or a user defined color. The function
decides whether or not to proceed with the clear depending on if there is
something pending to be drawn on the screen with the needRefresh boolean. The
default clear color has been changed from pink (255, 0, 255) to black
(0, 0, 0).