Game, GameState and Text classes have been enhanced, added Timer class.
A major addition to the Game class is a means to limiting the number of frames
per second rendered while the game loop is running. A max frame rate and capped
frame rate is possible with the default frame rate capped at 30. Constructor
for the Game class has been optimized with an initializer list. SDL_Event
is no longer created once per loop but once per Game instance for better
performance. GameState no longer copies the event object, but takes a
reference to it. Timer class has been added to the library for keeping track
of time in a stopwatch manner. And Hume header adds the Timer class.