Updated Image class.
The Image class has been updated to support SDL2. Color keying the
image now works slightly different since SDL changes how drawing
to the screen works. The color key can only be set BEFORE an image
is loaded. Otherwise, setting the keying color is in vain and will
not work unless the image is reloaded. loadImage now requires the
working SDL_Renderer be passed to it as well now.