Updated Window class.
Window class is now ready for SDL2. A lot of functions have been
reworked and removed from the class to allow it to be complete. There
is a really kludgy implementation for knowing the width and height of the
window object. As of now, it is only set when the constructor is used for
the object and will not take resizing into account, which is potentially
dangerous.

Also, the Quick-Align functions have been removed until there is a way
to retrieve the width and height of Blittable objects given to the Window.
They will be back in a later commit.