Prevent crash due to no GameState present.
During the update portion of the loop. It is possible for the state to
pop itself from the stack. Then, in the display function, the Game object
requests that the GameState draw its entities to the window. If the state
was popped from the stack during update(), the application would crash.