Completed basic implementation of Tile and TileMap class.
The Tile and TileMap classes are not ready to be used
experimentally. There is still a lot of work to be done with
the library in regards to its architecture. The Tile class has
been changed to inherit from Blittable, as it is more appropriate
to that.

The TileMap class keeps track of tiles as a matrix of pointers. This
may be changed later to actual objects with pointers to their tile type.

The Window class also has a function to draw an ENTIRE map to the screen
and not just a specific portion to a specific part. So larger tiled maps could be
a problem on constrained systems.