Center maps smaller than window.
Added some code to center maps on the screen that are smaller than
the window they are displayed in. The x_start and y_start variables
denote the offset needed for drawing and are added and subtracted
when obtaining other values such as during retrieval of tiles in the
matrix.