Beginning integration of new Sprite class.
The Sprite class is now being worked in as a subclass of Image and
has brought along quite a few changes for all three of them along with
the Window class.

The Blittable class' getWidth() and getHeight() functions are defined
virtual to allow Sprite to return individual sprite sizes upon request
(getDimensions() may stay the same though).

The Image class' loadImage() function is virtual as well so Sprite can
update the dimensions of its sprites upon finishing the loading of an image.

Major overhaul to the Sprite class from the last commit regarding it. I'm still
currently working it into the library and a lot of code has been removed to be
reserved for specializations in specific games.

And, finally, the Window class has a function to draw Sprites to the
screen. The selection of which is dependent upon the class.