Add scripts for obtaining test resources.
Two new scripts are available. The first, 'check-resources', can
look at the working directory and see if the files within are
sufficient for running the unit tests.

The second, 'get-resources', is an easy way of downloading a zip
file and extracting it into the working directory.

The makefile will now run check-resources, refusing to contine
with unit testing if the script does not find what is necessary.